+++ src/index.ts
@@ -90,6 +90,8 @@ export {
   TabPanels,
   TabPanel,
 } from "./components/Tabs/Tabs";
+export { default as Text } from "./components/Text/Text";
+export { TextDisplaySizes } from "./components/Text/TextTypes";
 export {
   TemplateAppContainer,
   Template,
@@ -101,6 +103,7 @@ export {
   TemplateContentSidebar,
   TemplateFooter,
 } from "./components/Template/Template";
+
 export {
   default as TextInput,
   TextInputRefType,