+++ src/theme/components/link.ts
@@ -3,13 +3,19 @@ const variants = {
   moreLink: {
     alignItems: "center",
     display: "inline-flex",
+    svg: {
+      height: "s",
+      width: "s",
+      textDecoration: "none",
+      fill: "currentColor",
+    },
     _hover: {
       textDecoration: "underline",
     },
   },
   button: {
     width: "100px",
-    borderRadius: "2px",
+    borderRadius: "sm",
     lineHeight: "1.5",
     display: "flex",
     cursor: "pointer",