+++ src/theme/components/textInput.ts
@@ -3,7 +3,7 @@ import { activeFocus, helperTextMargin } from "./global";
 const input = {
   border: "1px solid",
   borderColor: "ui.gray.medium",
-  borderRadius: "2px",
+  borderRadius: "sm",
   py: "xs",
   px: "s",
   _hover: {
@@ -31,6 +31,7 @@ const input = {
 const TextInput = {
   parts: ["helper", "input", "textarea"],
   baseStyle: {
+    marginBottom: "xs",
     helper: {
       ...helperTextMargin,
     },
@@ -39,6 +40,20 @@ const TextInput = {
       ...input,
       lineheight: "1.5",
       minHeight: "xxl",
+    },
+  },
+  variants: {
+    searchBar: {
+      flex: "1 1 80%",
+      input: {
+        borderRightRadius: "none",
+      },
+    },
+    searchBarSelect: {
+      flex: "1 1 80%",
+      input: {
+        borderRadius: "none",
+      },
     },
   },
 };