+++ CHANGELOG.md
@@ -10,8 +10,16 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 ### Adds
 
+- Adds custom `useCarouselStyles` hook for the `Tabs` mobile functionality.
+- Adds custom NYPL-theme radii values to use within Chakra.
+
+## 0.25.1 (October 14, 2021)
+
+### Adds
+
 - Updates the `Accordion` component to internally use Chakra components.
 - Adds DS `SimpleGrid` component composed from Chakra's `Simplegrid` component.
+- Adds DS `Text` component composed from Chakra's `Text` component.
 - Update the NYPL Colors with updated color names and values.
 - Composes the DS `TextInput` component from Chakra's `Input` and `Textarea` components.
 - Composes the DS `Select` component from Chakra's `Select` component.
@@ -21,12 +29,20 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 - Composes the DS `TemplateAppContainer` component from Chakra's `Box` component for page layouts through props.
 - Composes the DS `Link` component with Chakra component and fixes the `Button` type.
 - Exports Chakra's `Grid` and `GridItem` components.
+- Composes the DS `Label` component with Chakra component and updates the font size.
+- Composes the DS `StatusBadge` component from Chakra's component.
+- Composes the DS `Forms` component using DS and Chakra components.
+- Composes the DS `Icon` component with Chakra's `Icon` component. Added new props for `Icon` sizing, alignment, name, and rotation.
+- Composes the DS `SearchBar` component from DS `Select`, `TextInput`, `Button`, and `HelperErrorText` components.
+- Composes the DS `Breadcrumbs` component from Chakra's `Breadcumb`, `BreadcrumbItem` and `BreadcrumbLink` components.
 
 ### Changes
 
-- Updating spacing values to use NYPL spacing rather than "em"s.
+- Updates spacing values to use NYPL spacing rather than "em"s.
 - Updates the `SkeletonLoaderLayouts` enum values from `Horizontal` and `Vertical` to `Row` and `Column`, respectively.
 - Updates the `CardLayouts` enum values from `Horizontal` and `Vertical` to `Row` and `Column`, respectively.
+- Updates font size of the `legend` element used in `CheckboxGroup`, `DatePicker` and `RadioGroup`.
+- Updates the radius values through Chakra's theme object.
 
 ## 0.25.0